Moves annotation list filtering to Atlas component
Moving the filtering up to this level, means we don't have to pass all
the filter variables up from the filter controls and back down to the
annotations list. It also makes it possible to restore the correct page
selection when moving from showing 4 items per page to 12 and back.
We load the annotation data in this component, so it is not unreasonable
to filter them here too.