Isolates FocusedProofreading & OrphanLink neuroglancer
This change isolates the neuroglancer component in the
FocusedProofreading & OrphanLink components, to prevent them from being
modified by another component from another workspace, such as the
Annotation component. It also leaves neuroglancer in a "default",
unmodified state, which means it will need to be modified by loading in
information from the FocusedProofreading or OrphanLink job files.