fix: Restores auth token for use by neuroglancer
The Neuroglancer code was using the 'neurohub' key on window to get the
auth token that it needed in order for the clio datasource to access the
data in clio store. This was removed when we switched over from google
tokens to flyem tokens. Adding this back fixes the issues that people
were seeing when trying to access annotations in the annotate workspace.