fix: state preservation across reload for Annotate
The state preservation was not working, because the state that was saved
in sessionStorage had a different key to the one being requested when
the page reloaded. This was tracked down to an incorrect title being set
on the neuroglancer object that was stored in the clio datastore. To fix
this, the code overwrites any titles that are found in the stored
neuroglancer objects with the dataset 'key', this will allow people to
continue to access their stored state, even if the version number is
updated.