fix: session storage key reverted to dataset-version key
Having the sessionStorage key as just the dataset caused too many querks
with the way data was loaded into the browser after a version change.
The user would see the new version in the dataset selection dropdown,
but would still be using old layers from the session storage as no
merging is performed on the dataset information retrieved from clio and
the data retrieved from the sessionStorage. Having the session removed
on a dataset version change is safer for everyone.