fix: Upgrades the login button to use fedCM
There were a few changes to the login button to get the fedCM login
prompt to appear correctly in Chrome.

From google:

"On January 4, 2024, Chrome began restricting access to third-party
cookies by default for one percent of Chrome browsers globally. This was
announced in May of 2023 and the timeline was confirmed in the Privacy
Sandbox team’s blog post on December 14, 2023. This one percent group is
intended to support industry testing of the Privacy Sandbox
technologies, as explained in the Chrome-facilitated testing guidance
and the UK Competition and Market Authority’s guidance to third parties
on testing.

You are receiving this message because one or more of your web
applications uses the One Tap - Sign In with Google for Web SDK which
will be migrating to Federated Credential Management (FedCM) in April
2024. Verify whether changes to your website are necessary using the
instructions below."