fix: Better dvid error reporting for annotations
The error reporting code was not showing the end user the error message
that was being returned by the dvid server, because it was not parsing
the json payload in the response if non 200 status was returned. This
update adds the json parsing of the response body to the error handling,
so that the error message can be parsed and displayed to the user.