fix: Removes replace=true from updateBodyAnnotation
It is believed that the updateBodyAnnotation function is replacing
annotations with incomplete information, instead of adding the
information to the existing data. This change removes the replace=true
from the updateBodyAnnotation to prevent the overwriting from happening.
The Body Status select box in the Orphanlink component was also
disabled, until we can figure out if it is needed and how this change
will affect its' functionality.