fix: Body annotation fields return null when deleted.
Prior to this, the annotation fields were set to undefined when deleted.
This was ignored by the backend API, unless the replace=true flag was
set. Since replace=true is dangerous, we have replaced it with
replace=false and therefore need to make sure that the null values are
passed, so that people can delete annotations.