fix: Body annotations removed from POST when null and previously undefined
The fix that was introduced in commit 0e5a1b13cfb caused the backed to be
overzealous when updating the database as all fields received a null
value, even if they were not edited. This forced the backend to update
the user and timestamp values for all fields that were not touched when
an update occurred. The fix here is to remove those fields from the
update that have a null value and were previously undefined. That way
they never get passed to the backend and the fields don't get updated.